Rename test srcs and drop compiler directives
