+++ tests/core/go_binary/pie_linux_test.go
@@ -1,9 +1,7 @@
-// +build darwin
-
 package test
 
 import (
-	"debug/macho"
+	"debug/elf"
 	"fmt"
 	"os"
 	"testing"
@@ -11,7 +9,7 @@ import (
 	"github.com/bazelbuild/rules_go/go/tools/bazel"
 )
 
-func openMachO(dir, bin string) (*macho.File, error) {
+func openELF(dir, bin string) (*elf.File, error) {
 	bin, ok := bazel.FindBinary(dir, bin)
 	if !ok {
 		return nil, fmt.Errorf("could not find binary: %s", bin)
@@ -22,27 +20,29 @@ func openMachO(dir, bin string) (*macho.File, error) {
 		return nil, err
 	}
 
-	return macho.NewFile(f)
+	return elf.NewFile(f)
 }
 
 func TestPIE(t *testing.T) {
-	m, err := openMachO("tests/core/go_binary", "hello_pie_bin")
+	e, err := openELF("tests/core/go_binary", "hello_pie_bin")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if m.Flags&macho.FlagPIE == 0 {
+	// PIE binaries are implemented as shared libraries.
+	if e.Type != elf.ET_DYN {
 		t.Error("ELF binary is not position-independent.")
 	}
 }
 
 func TestNoPIE(t *testing.T) {
-	m, err := openMachO("tests/core/go_binary", "hello_nopie_bin")
+	e, err := openELF("tests/core/go_binary", "hello_nopie_bin")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if m.Flags&macho.FlagPIE != 0 {
+	// PIE binaries are implemented as shared libraries.
+	if e.Type != elf.ET_EXEC {
 		t.Error("ELF binary is not position-dependent.")
 	}
 }
